// src/db/generated/models/PayrollInputs.ts
export type PayrollInputsModel = runtime.Types.Result.DefaultSelection<Prisma.$PayrollInputsPayload>
// ...
export type PayrollInputsSelectScalar = {
  id?: boolean
  agencyId?: boolean
  caregiverId?: boolean
  hourlyRateCents?: boolean
  overtimeRateCents?: boolean
  payFrequency?: boolean
  w4FilingStatus?: boolean
  w4MultipleJobs?: boolean
  w4DependentsAmountCents?: boolean
  w4OtherIncomeCents?: boolean
  w4DeductionsCents?: boolean
  w4ExtraWithholdingCents?: boolean
  it2104FilingStatus?: boolean
  it2104AllowancesNy?: boolean
  it2104AllowancesNyc?: boolean
  it2104AllowancesYonkers?: boolean
  it2104ExtraWithholdingNyCents?: boolean
  it2104ResidentNyc?: boolean
  it2104ResidentYonkers?: boolean
  bankName?: boolean
  bankAccountType?: boolean
  bankRoutingNumberEnc?: boolean
  bankAccountNumberEnc?: boolean
  bankAccountLast4?: boolean
  directDepositAuthorizedAt?: boolean
  createdAt?: boolean
  updatedAt?: boolean
}
// ...
export type $PayrollInputsPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "PayrollInputs"
  objects: {
    caregiver: Prisma.$CaregiverPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    caregiverId: string
    hourlyRateCents: number | null
    overtimeRateCents: number | null
    payFrequency: string | null
    w4FilingStatus: string | null
    w4MultipleJobs: boolean | null
    w4DependentsAmountCents: number | null
    w4OtherIncomeCents: number | null
    w4DeductionsCents: number | null
    w4ExtraWithholdingCents: number | null
    it2104FilingStatus: string | null
    it2104AllowancesNy: number | null
    it2104AllowancesNyc: number | null
    it2104AllowancesYonkers: number | null
    it2104ExtraWithholdingNyCents: number | null
    it2104ResidentNyc: boolean | null
    it2104ResidentYonkers: boolean | null
    bankName: string | null
    bankAccountType: string | null
    /**
     * @tier sensitive @encrypted
     */
    bankRoutingNumberEnc: runtime.Bytes | null
    /**
     * @tier sensitive @encrypted
     */
    bankAccountNumberEnc: runtime.Bytes | null
    bankAccountLast4: string | null
    directDepositAuthorizedAt: Date | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["payrollInputs"]>
  composites: {}
}